// resources/js/game/sections/components/gems/deffinitions/attached-gems.ts
export default interface AttachedGems {
    id: number;
    name: string;
    primary_atonement_amount: number;
    primary_atonement_name: string;
    secondary_atonement_amount: number;
    secondary_atonement_name: string;
    tertiary_atonement_amount: number;
    tertiary_atonement_name: string;
    tier: number;
}